Migrated main score header layout to encompass both games and menu
